Modified BECCS liquid fuels assets
Previously, the BECCS liquid fuels go directly to the respective fuels end use demand nodes, but it is better to treat each respective fuel nodes (not the demand node but the node with price to purchase fossil fuel) as a blending node for purchased fossil fuel, biofuels and synthetic fuels, before their respective end use transformations to the demand node. This tidy things up as the end use transformation that accounts for fuel combustion emissions are only for the end use demand node.